requirements.txt
chromadb
sentence-transformers
python-dotenv
groq
langchain
pypdf
termcolor